Rename quality selection radio buttons component
